+++ frontend/src/components/Layout/TheFooter.vue
@@ -1,5 +1,5 @@
 <template>
-    <footer class="flex justify-center content-center px-4 py-8 bottom-0 bg-teal-800 dark:bg-gray-900">
+    <footer class="flex justify-center content-center px-4 py-8 bottom-0 bg-teal-800 dark:bg-zinc-900">
         <p class="text-gray-100">&copy; 2023 Copyright Patrick Geiger</p>
     </footer>
 </template>