+++ frontend/src/router/router.ts
@@ -1,28 +1,29 @@
-import { createRouter, createWebHistory } from "vue-router"
-import MainPage from '../view/MainPage.vue'
-import ResultsPage from '../view/ResultsPage.vue'
-import RequestPage from '../view/RequestPage.vue'
-import SearchPage from '../view/SearchPage.vue'
+import { createRouter, createWebHistory } from "vue-router";
+import MainPage from "../view/MainPage.vue";
+import ResultsPage from "../view/ResultsPage.vue";
+import SearchPage from "../view/SearchPage.vue";
+import AboutPage from "../view/AboutPage.vue";
 
 const router = createRouter({
     history: createWebHistory(),
     routes: [
         {
-            path: '/',
-            component: MainPage
-        },{
-            path: '/request',
-            component: RequestPage
-        }, 
-        {
-            path: '/results',
-            component: ResultsPage
+            path: "/",
+            component: MainPage,
         },
         {
-            path: '/search',
-            component: SearchPage
-        }
-    ]
+            path: "/about",
+            component: AboutPage,
+        },
+        {
+            path: "/results",
+            component: ResultsPage,
+        },
+        {
+            path: "/search",
+            component: SearchPage,
+        },
+    ],
 });
 
 export default router;