+++ frontend/src/view/MainPage.vue
@@ -1,19 +1,23 @@
 <template>
-    <main class="flex justify-center">
-        <BasicCard class="w-1/2 text-center text-gray-900 dark:text-gray-200">
-            <h1 class="text-2xl mb-4">
+    <main class="flex justify-end content-center flex-col m-16">
+        <BasicCard class="text-center text-gray-900 dark:text-gray-200 w-full lg:3/5 m-auto">
+            <h1 class="text-4xl mb-4 font-bold">
                 Green Web Analyzer
             </h1>
-            <p class="w-3/5 m-auto">
+            <p class="w-4/5 m-auto text-justify">
                 This website is a tool that allows you to check a website for certain criterias. With the background of
                 Green IT, websites are to be checked with regard to data minimisation and efficiency.
             </p>
+            <RequestForm />
         </BasicCard>
+
+
     </main>
 </template>
 
 <script lang="ts">
 import { defineComponent } from 'vue'
+import RequestForm from '../components/RequestForm/RequestForm.vue';
 
-export default defineComponent({})
+export default defineComponent({ components: { RequestForm } })
 </script>