+++ CHANGELOG.md
@@ -1,1 +1,15 @@
-# [0.1.0]
+# 0.2.2
+
+## Added
+
+-   Added RequestForm to MainPage
+-   Added basic AboutPage
+
+## Updated
+
+-   Updated dark theme
+-   Updated CriteriaRequest to handle redirects and requests with no responses
+
+## Removed
+
+-   Remove extra Page for Requests